add multisig address for 5% proposal allocation
Based on the community discussion detailed in OIP-10, this commit adds 5% of the total supply to a single multisig wallet in the genesis file, replacing 2 previous addresses that contained placeholder amounts. This finalizes the wallet setup with the correct, agreed-upon amount on a single multisig wallet address as per the community's decision.